Move JSONPath eval to eventfilters
