+++ backend/pkg/services/alerts/event_filters/jsonpathfilter.go
@@ -2,13 +2,22 @@ package eventfilters
 
 import (
 	cloudevents "github.com/cloudevents/sdk-go/v2"
-	"github.com/lamassuiot/lamassuiot/backend/v3/pkg/helpers"
 	"github.com/lamassuiot/lamassuiot/core/v3/pkg/models"
+	"github.com/spyzhov/ajson"
 )
 
 func RegisterJSONPathFilter() {
 	RegisterEventFilter(models.JSONPath,
 		func(c models.SubscriptionCondition, event cloudevents.Event) (bool, error) {
-			return helpers.JsonPathExists(event.Data(), c.Condition)
+			return JsonPathExists(event.Data(), c.Condition)
 		})
+}
+
+func JsonPathExists(data []byte, path string) (bool, error) {
+	ajsonData, err := ajson.JSONPath(data, path)
+	if err != nil {
+		return false, err
+	}
+
+	return len(ajsonData) > 0, nil
 }
