+++ backend/pkg/services/alerts/event_filters/jsonpathfilter_test.go
@@ -1,4 +1,4 @@
-package helpers
+package eventfilters
 
 import (
 	"encoding/json"
